requirements.txt
# Core ML & Quantum
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
pennylane==0.35.1
pennylane-lightning==0.35.1

# NLP & Transformers
transformers==4.46.2
sentence-transformers==3.0.1
datasets==2.21.0

# Data & Scientific
numpy==1.26.4
pandas==2.2.0
scipy==1.13.1
scikit-learn==1.4.2

# Utilities
PyYAML==6.0.1
tqdm==4.66.4
joblib==1.4.2
psutil==6.0.0
rich==13.7.1

# Logging & Config
pydantic==2.9.1
pydantic-settings==2.6.1

# Visualization (Optional)
matplotlib==3.8.4
seaborn==0.13.2
